Cleaned and made the play.py the script to use for playing
