+++ pyproject.toml
@@ -12,7 +12,6 @@ dependencies = [
     "ml-dtypes==0.5.0",
     "numpy==2.1.3",
     "opt-einsum==3.4.0",
-    "pygame==2.6.1",
     "scipy==1.14.1",
     "toolz==1.0.0",
     "typing-extensions==4.12.2",
