+++ src/jaxatari/games/jax_seaquest.py
@@ -4,9 +4,7 @@ from typing import Tuple, NamedTuple
 import jax
 import jax.numpy as jnp
 import chex
-import pygame
 import jaxatari.rendering.atraJaxis as aj
-import numpy as np
 import jaxatari.spaces as spaces
 
 from jaxatari.environment import JaxEnvironment, JAXAtariAction as Action
@@ -3224,113 +3222,3 @@ class SeaquestRenderer(AtraJaxisRenderer):
         raster = raster.at[0:bar_width, :, :].set(0)
 
         return raster
-
-
-def get_human_action() -> chex.Array:
-    """Get human action from keyboard with support for diagonal movement and combined fire"""
-    keys = pygame.key.get_pressed()
-    up = keys[pygame.K_UP] or keys[pygame.K_w]
-    down = keys[pygame.K_DOWN] or keys[pygame.K_s]
-    left = keys[pygame.K_LEFT] or keys[pygame.K_a]
-    right = keys[pygame.K_RIGHT] or keys[pygame.K_d]
-    fire = keys[pygame.K_SPACE]
-
-    # Diagonal movements with fire
-    if up and right and fire:
-        return jnp.array(Action.UPRIGHTFIRE)
-    if up and left and fire:
-        return jnp.array(Action.UPLEFTFIRE)
-    if down and right and fire:
-        return jnp.array(Action.DOWNRIGHTFIRE)
-    if down and left and fire:
-        return jnp.array(Action.DOWNLEFTFIRE)
-
-    # Cardinal directions with fire
-    if up and fire:
-        return jnp.array(Action.UPFIRE)
-    if down and fire:
-        return jnp.array(Action.DOWNFIRE)
-    if left and fire:
-        return jnp.array(Action.LEFTFIRE)
-    if right and fire:
-        return jnp.array(Action.RIGHTFIRE)
-
-    # Diagonal movements
-    if up and right:
-        return jnp.array(Action.UPRIGHT)
-    if up and left:
-        return jnp.array(Action.UPLEFT)
-    if down and right:
-        return jnp.array(Action.DOWNRIGHT)
-    if down and left:
-        return jnp.array(Action.DOWNLEFT)
-
-    # Cardinal directions
-    if up:
-        return jnp.array(Action.UP)
-    if down:
-        return jnp.array(Action.DOWN)
-    if left:
-        return jnp.array(Action.LEFT)
-    if right:
-        return jnp.array(Action.RIGHT)
-    if fire:
-        return jnp.array(Action.FIRE)
-
-    return jnp.array(Action.NOOP)
-
-
-if __name__ == "__main__":
-    # Initialize game and renderer
-    game = JaxSeaquest()
-    pygame.init()
-    screen = pygame.display.set_mode((WIDTH * SCALING_FACTOR, HEIGHT * SCALING_FACTOR))
-    clock = pygame.time.Clock()
-
-    renderer_AtraJaxis = SeaquestRenderer()
-
-    # Get jitted functions
-    jitted_step = jax.jit(game.step)
-    jitted_reset = jax.jit(game.reset)
-
-    curr_obs, curr_state = jitted_reset()
-
-    # Game loop with rendering
-    running = True
-    frame_by_frame = False
-    frameskip = 1
-    counter = 1
-
-    while running:
-        for event in pygame.event.get():
-            if event.type == pygame.QUIT:
-                running = False
-            elif event.type == pygame.KEYDOWN:
-                if event.key == pygame.K_f:
-                    frame_by_frame = not frame_by_frame
-            elif event.type == pygame.KEYDOWN or (
-                event.type == pygame.KEYUP and event.key == pygame.K_n
-            ):
-                if event.key == pygame.K_n and frame_by_frame:
-                    if counter % frameskip == 0:
-                        action = get_human_action()
-                        curr_obs, curr_state, reward, done, info = jitted_step(
-                            curr_state, action
-                        )
-                        print(f"Observations: {curr_obs}")
-                        print(f"Reward: {reward}, Done: {done}, Info: {info}")
-
-        if not frame_by_frame:
-            if counter % frameskip == 0:
-                action = get_human_action()
-                curr_obs, curr_state, reward, done, info = jitted_step(
-                    curr_state, action
-                )
-
-        # render and update pygame
-        raster = renderer_AtraJaxis.render(curr_state)
-        aj.update_pygame(screen, raster, SCALING_FACTOR, WIDTH, HEIGHT)
-        counter += 1
-        clock.tick(60)
-
-    pygame.quit()
